Adds protection against non-authorized user navigating to delete page via url

+++ project.py
@@ -175,11 +175,10 @@ def restaurantsJSON():
 def showRestaurants():
   restaurants = session.query(Restaurant).order_by(asc(Restaurant.name))
 
-  try:
-      currentUser = getUserID(login_session['email'])
-      return render_template('restaurants.html', restaurants = restaurants)
-  except:
+  if 'username' not in login_session:
       return render_template('publicrestaurants.html', restaurants = restaurants)
+  else:
+      return render_template('restaurants.html', restaurants = restaurants)
 
 
 #Create a new restaurant
@@ -216,7 +215,12 @@ def editRestaurant(restaurant_id):
 def deleteRestaurant(restaurant_id):
     if 'username' not in login_session:
         return redirect('/login')
+
     restaurantToDelete = session.query(Restaurant).filter_by(id = restaurant_id).one()
+    if restaurantToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized \
+        to delete this restaurant. Please create your own restaurant in order to delete.');\
+        }</script><body onload='myFunction()''>"
     if request.method == 'POST':
         session.delete(restaurantToDelete)
         flash('%s Successfully Deleted' % restaurantToDelete.name)
@@ -231,14 +235,13 @@ def deleteRestaurant(restaurant_id):
 def showMenu(restaurant_id):
     restaurant = session.query(Restaurant).filter_by(id = restaurant_id).one()
     items = session.query(MenuItem).filter_by(restaurant_id = restaurant_id).all()
+
     creator = getUserInfo(restaurant.user_id)
 
-    currentUser = getUserID(login_session['email'])
-
-    if currentUser == creator.id:
-        return render_template('menu.html', items = items, restaurant = restaurant, creator = creator)
-    else:
+    if 'username' not in login_session or creator.id != login_session['user_id']:
         return render_template('publicmenu.html', items = items, restaurant = restaurant, creator = creator)
+    else:
+        return render_template('menu.html', items = items, restaurant = restaurant, creator = creator)
